+++ changelog.md
@@ -33,6 +33,7 @@
 * LLT-4013: Add conntrack rules for icmp
 * LLT-4152: Suspend SessionKeeper and CrossPingCheck for unresponsive peers
 * LLT-4360: Bump rust version to 1.72
+* LLT-4286: Add icmp error packet handling to firewall
 
 <br>